requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Web scraping
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
lxml==4.9.3
aiohttp==3.9.1
playwright==1.40.0

# Task scheduling
celery==5.3.4
redis==5.0.1
apscheduler==3.10.4

# Data processing
pandas==2.1.4
numpy==1.25.2

# Environment and configuration
python-dotenv==1.0.0
python-multipart==0.0.6

# Logging and monitoring
structlog==23.2.0
sentry-sdk[fastapi]==1.38.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
pytest-cov==4.1.0

# Development tools
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.3.0
bcrypt==4.0.1

# OpenAI
openai==1.40.0